requirements.txt
# Web Framework
fastapi
uvicorn

# Vector search
faiss-cpu

# Embeddings
sentence-transformers

# Hugging Face Models
transformers>=4.36.0

torch

# utilities
numpy
python-dotenv

# for ui
streamlit

